+++ dist2/src/data-model/excel-table.js
@@ -1,4 +1,3 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-// Now you can use 'YourObject' as the type for your data.
 //# sourceMappingURL=excel-table.js.map