+++ dist2/src/index.js
@@ -42,157 +42,10 @@ const create_excel_data_1 = require("./utils/create-excel-data");
 const color_1 = require("./utils/color");
 const comment_1 = require("./utils/comment");
 const multi_value_1 = require("./utils/multi-value");
+const const_data_1 = require("./utils/content-generator/const-data");
 function generateExcel(data) {
     return __awaiter(this, void 0, void 0, function* () {
-        let formatMap = {
-            time: {
-                key: 165,
-                value: '<numFmt numFmtId="165" formatCode="[$-F400]h:mm:ss\\ AM/PM" />',
-            },
-            date: {
-                key: 187,
-                value: '<numFmt numFmtId="187" formatCode="[$-F800]dddd\\,\\ mmmm\\ dd\\,\\ yyyy" />',
-            },
-            short_date: {
-                key: 14,
-            },
-            fraction: {
-                key: 13,
-            },
-            percentage: {
-                key: 9,
-            },
-            float_1: { key: 180, value: '<numFmt numFmtId="180" formatCode="0.0" />' },
-            float_2: { key: 181, value: '<numFmt numFmtId="181" formatCode="0.00" />' },
-            float_3: {
-                key: 164,
-                value: '<numFmt numFmtId="164" formatCode="0.000" />',
-            },
-            float_4: {
-                key: 182,
-                value: '<numFmt numFmtId="182" formatCode="0.0000" />',
-            },
-            dollar_2: {
-                key: 183,
-                value: '<numFmt numFmtId="183" formatCode="&quot;$&quot;#,##0.00" />',
-            },
-            num_sep: {
-                key: 184,
-                value: '<numFmt numFmtId="184" formatCode="#,##0" />',
-            },
-            num_sep_1: {
-                key: 185,
-                value: '<numFmt numFmtId="185" formatCode="#,##0.0" />',
-            },
-            num_sep_2: {
-                key: 186,
-                value: '<numFmt numFmtId="186" formatCode="#,##0.00" />',
-            },
-            dollar: {
-                key: 163,
-                value: '<numFmt numFmtId="163" formatCode="_([$$-409]* #,##0.00_);_([$$-409]* \\(#,##0.00\\);_([$$-409]* &quot;-&quot;??_);_(@_)" />',
-            },
-            $: {
-                key: 163,
-                value: '<numFmt numFmtId="163" formatCode="_([$$-409]* #,##0.00_);_([$$-409]* \\(#,##0.00\\);_([$$-409]* &quot;-&quot;??_);_(@_)" />',
-            },
-            pound: {
-                key: 162,
-                value: '<numFmt numFmtId="162" formatCode="_-[$£-809]* #,##0.00_-;\\-[$£-809]* #,##0.00_-;_-[$£-809]* &quot;-&quot;??_-;_-@_-" />',
-            },
-            "£": {
-                key: 162,
-                value: '<numFmt numFmtId="162" formatCode="_-[$£-809]* #,##0.00_-;\\-[$£-809]* #,##0.00_-;_-[$£-809]* &quot;-&quot;??_-;_-@_-" />',
-            },
-            euro: {
-                key: 161,
-                value: '<numFmt numFmtId="161" formatCode="_([$€-2]\\ * #,##0.00_);_([$€-2]\\ * \\(#,##0.00\\);_([$€-2]\\ * &quot;-&quot;??_);_(@_)" />',
-            },
-            "€": {
-                key: 161,
-                value: '<numFmt numFmtId="161" formatCode="_([$€-2]\\ * #,##0.00_);_([$€-2]\\ * \\(#,##0.00\\);_([$€-2]\\ * &quot;-&quot;??_);_(@_)" />',
-            },
-            yen: {
-                key: 160,
-                value: '<numFmt numFmtId="160" formatCode="_ [$¥-804]* #,##0.00_ ;_ [$¥-804]* \\-#,##0.00_ ;_ [$¥-804]* &quot;-&quot;??_ ;_ @_ " />',
-            },
-            "¥": {
-                key: 160,
-                value: '<numFmt numFmtId="160" formatCode="_ [$¥-804]* #,##0.00_ ;_ [$¥-804]* \\-#,##0.00_ ;_ [$¥-804]* &quot;-&quot;??_ ;_ @_ " />',
-            },
-            CHF: {
-                key: 179,
-                value: '<numFmt numFmtId="179" formatCode="_-* #,##0.00\\ [$CHF-100C]_-;\\-* #,##0.00\\ [$CHF-100C]_-;_-* &quot;-&quot;??\\ [$CHF-100C]_-;_-@_-" />',
-            },
-            ruble: {
-                key: 178,
-                value: '<numFmt numFmtId="178" formatCode="_-* #,##0.00\\ [$₽-419]_-;\\-* #,##0.00\\ [$₽-419]_-;_-* &quot;-&quot;??\\ [$₽-419]_-;_-@_-" />',
-            },
-            "₽": {
-                key: 178,
-                value: '<numFmt numFmtId="178" formatCode="_-* #,##0.00\\ [$₽-419]_-;\\-* #,##0.00\\ [$₽-419]_-;_-* &quot;-&quot;??\\ [$₽-419]_-;_-@_-" />',
-            },
-            "֏": {
-                key: 177,
-                value: '<numFmt numFmtId="177" formatCode="_-* #,##0.00\\ [$֏-42B]_-;\\-* #,##0.00\\ [$֏-42B]_-;_-* &quot;-&quot;??\\ [$֏-42B]_-;_-@_-" />',
-            },
-            manat: {
-                key: 176,
-                value: '<numFmt numFmtId="176" formatCode="_-* #,##0.00\\ [$₼-82C]_-;\\-* #,##0.00\\ [$₼-82C]_-;_-* &quot;-&quot;??\\ [$₼-82C]_-;_-@_-" />',
-            },
-            "₼": {
-                key: 176,
-                value: '<numFmt numFmtId="176" formatCode="_-* #,##0.00\\ [$₼-82C]_-;\\-* #,##0.00\\ [$₼-82C]_-;_-* &quot;-&quot;??\\ [$₼-82C]_-;_-@_-" />',
-            },
-            "₼1": {
-                key: 175,
-                value: '<numFmt numFmtId="175" formatCode="_-* #,##0.00\\ [$₼-42C]_-;\\-* #,##0.00\\ [$₼-42C]_-;_-* &quot;-&quot;??\\ [$₼-42C]_-;_-@_-" />',
-            },
-            "₽1": {
-                key: 174,
-                value: '<numFmt numFmtId="174" formatCode="_-* #,##0.00\\ [$₽-46D]_-;\\-* #,##0.00\\ [$₽-46D]_-;_-* &quot;-&quot;??\\ [$₽-46D]_-;_-@_-" />',
-            },
-            "₽2": {
-                key: 173,
-                value: '<numFmt numFmtId="173" formatCode="_-* #,##0.00\\ [$₽-485]_-;\\-* #,##0.00\\ [$₽-485]_-;_-* &quot;-&quot;??\\ [$₽-485]_-;_-@_-" />',
-            },
-            "₽3": {
-                key: 172,
-                value: '<numFmt numFmtId="172" formatCode="_-* #,##0.00\\ [$₽-444]_-;\\-* #,##0.00\\ [$₽-444]_-;_-* &quot;-&quot;??\\ [$₽-444]_-;_-@_-" />',
-            },
-            ريال: {
-                key: 171,
-                value: '<numFmt numFmtId="171" formatCode="_ * #,##0.00_-[$ريال-429]_ ;_ * #,##0.00\\-[$ريال-429]_ ;_ * &quot;-&quot;??_-[$ريال-429]_ ;_ @_ " />',
-            },
-        };
-        let cols = [
-            "A",
-            "B",
-            "C",
-            "D",
-            "E",
-            "F",
-            "G",
-            "H",
-            "I",
-            "J",
-            "K",
-            "L",
-            "M",
-            "N",
-            "O",
-            "P",
-            "Q",
-            "R",
-            "S",
-            "T",
-            "U",
-            "V",
-            "W",
-            "X",
-            "Y",
-            "Z",
-        ];
+        let cols = [...const_data_1.cols];
         if (data.numberOfColumn && data.numberOfColumn > 25) {
             cols = (0, generate_column_name_1.generateColumnName)(cols, data.numberOfColumn);
         }
@@ -344,7 +197,7 @@ function generateExcel(data) {
                     "<border>" + borderStr + "<diagonal />" + "</border>";
             }
             if (styl.format) {
-                const format = formatMap[styl.format];
+                const format = const_data_1.formatMap[styl.format];
                 if (format) {
                     indexes.formatIndex = format.key;
                     if ("value" in format) {
@@ -415,6 +268,8 @@ function generateExcel(data) {
             let sheetDataString = "";
             let sheetSizeString = "";
             let sheetSortFilter = "";
+            let mergesCellArray = Object.assign([], sheetData.merges);
+            let formulaSheetObj = Object.assign({}, sheetData.formula);
             let hasComment = false;
             let commentAuthor = [];
             let commentString = "";
@@ -438,10 +293,7 @@ function generateExcel(data) {
                         : "";
                     const tStyle = title.styleId ? title.styleId : "titleStyle";
                     const refString = cols[left] + "" + (rowCount + top);
-                    if (!sheetData.merges) {
-                        sheetData.merges = [];
-                    }
-                    sheetData.merges.push(refString +
+                    mergesCellArray.push(refString +
                         ":" +
                         cols[left + consommeCol - 1] +
                         (rowCount + consommeRow + top));
@@ -489,7 +341,7 @@ function generateExcel(data) {
                         sharedStringIndex++;
                         sharedStringMap[title.text] = title.text;
                         if (title.multiStyleValue) {
-                            sharedString += (0, multi_value_1.generateMultiStyleValue)(title.multiStyleValue, title.text, styleMapper.commentSintax.value);
+                            sharedString += (0, multi_value_1.generateMultiStyleValue)(title.multiStyleValue, title.text, styleMapper.commentSintax.value, tStyle);
                         }
                         else {
                             sharedString += "<si><t>" + title.text + "</t></si>";
@@ -568,10 +420,10 @@ function generateExcel(data) {
                         });
                         commentString += (0, comment_1.generateCommentTag)(refString, commentObj.commentStr, commentObj.commentStyl, authorId);
                     }
-                    const formula = sheetData.formula && sheetData.formula[refString];
+                    const formula = formulaSheetObj && formulaSheetObj[refString];
                     if (formula) {
                         sheetDataString += (0, generate_formula_cell_1.generateCellRowCol)(refString, formula, data.styles).cell;
-                        delete sheetData.formula[refString];
+                        delete formulaSheetObj[refString];
                     }
                     else {
                         sheetDataString +=
@@ -587,7 +439,7 @@ function generateExcel(data) {
                                 sharedStringIndex +
                                 "</v></c>";
                         if (v.multiStyleValue) {
-                            sharedString += (0, multi_value_1.generateMultiStyleValue)(v.multiStyleValue, v.text, styleMapper.commentSintax.value);
+                            sharedString += (0, multi_value_1.generateMultiStyleValue)(v.multiStyleValue, v.text, styleMapper.commentSintax.value, headerStyleKey ? headerStyleKey : "");
                         }
                         else {
                             sharedString += "<si><t>" + v.text + "</t></si>";
@@ -675,10 +527,7 @@ function generateExcel(data) {
                                                 (rowCount + mergeValue[0]);
                                     }
                                 }
-                                if (!sheetData.merges) {
-                                    sheetData.merges = [];
-                                }
-                                sheetData.merges.push(mergeStr);
+                                mergesCellArray.push(mergeStr);
                             }
                         }
                         const rowStyle = mData.rowStyle;
@@ -722,14 +571,7 @@ function generateExcel(data) {
                                 }
                                 else {
                                     if (item && item.inProgress) {
-                                        if (!sheetData.merges) {
-                                            sheetData.merges = [
-                                                columnKey + item.start + ":" + columnKey + (rowCount - 1),
-                                            ];
-                                        }
-                                        else {
-                                            sheetData.merges.push(columnKey + item.start + ":" + columnKey + (rowCount - 1));
-                                        }
+                                        mergesCellArray.push(columnKey + item.start + ":" + columnKey + (rowCount - 1));
                                         mergeRowConditionMap[columnKey] = {
                                             inProgress: false,
                                             start: -1,
@@ -774,10 +616,10 @@ function generateExcel(data) {
                                     }
                                     commentString += (0, comment_1.generateCommentTag)(refString, commentObj.commentStr, commentObj.commentStyl, authorId);
                                 }
-                                const formula = sheetData.formula && sheetData.formula[refString];
+                                const formula = formulaSheetObj && formulaSheetObj[refString];
                                 if (formula) {
                                     sheetDataString += (0, generate_formula_cell_1.generateCellRowCol)(refString, formula).cell;
-                                    delete sheetData.formula[refString];
+                                    delete formulaSheetObj[refString];
                                 }
                                 else {
                                     if (typeof dataEl == "string") {
@@ -795,7 +637,7 @@ function generateExcel(data) {
                                         if ("multiStyleValue" in mData &&
                                             mData.multiStyleValue &&
                                             key in mData.multiStyleValue) {
-                                            sharedString += (0, multi_value_1.generateMultiStyleValue)(mData.multiStyleValue[key], dataEl, styleMapper.commentSintax.value);
+                                            sharedString += (0, multi_value_1.generateMultiStyleValue)(mData.multiStyleValue[key], dataEl, styleMapper.commentSintax.value, cellStyle ? cellStyle : "");
                                         }
                                         else {
                                             sharedString += "<si><t>" + dataEl + "</t></si>";
@@ -822,14 +664,7 @@ function generateExcel(data) {
                         if (rowLength - 1 == innerIndex) {
                             Object.keys(mergeRowConditionMap).forEach((v) => {
                                 if (mergeRowConditionMap[v].inProgress) {
-                                    if (sheetData.merges) {
-                                        sheetData.merges.push(v + mergeRowConditionMap[v].start + ":" + v + rowCount);
-                                    }
-                                    else {
-                                        sheetData.merges = [
-                                            v + mergeRowConditionMap[v].start + ":" + v + rowCount,
-                                        ];
-                                    }
+                                    mergesCellArray.push(v + mergeRowConditionMap[v].start + ":" + v + rowCount);
                                 }
                             });
                         }
@@ -855,22 +690,19 @@ function generateExcel(data) {
                     }
                 }
                 if (headerFormula.length > 0) {
-                    if (!sheetData.formula) {
-                        sheetData.formula = {};
-                    }
                     headerFormula.forEach((v) => {
                         const header = sheetData.headers[v];
-                        sheetData.formula[cols[v] + "" + rowCount] = Object.assign({ start: sheetData.withoutHeader ? cols[v] + "1" : cols[v] + "2", end: cols[v] + "" + (rowCount - 1), type: header.formula.type }, (header.formula.styleId
+                        formulaSheetObj[cols[v] + "" + rowCount] = Object.assign({ start: sheetData.withoutHeader ? cols[v] + "1" : cols[v] + "2", end: cols[v] + "" + (rowCount - 1), type: header.formula.type }, (header.formula.styleId
                             ? { styleId: header.formula.styleId }
                             : {}));
                     });
                 }
-                if (sheetData.formula) {
-                    const remindFormulaKey = Object.keys(sheetData.formula);
+                if (formulaSheetObj) {
+                    const remindFormulaKey = Object.keys(formulaSheetObj);
                     if (remindFormulaKey.length) {
                         let rF = {};
                         remindFormulaKey.forEach((v) => {
-                            const f = (0, generate_formula_cell_1.generateCellRowCol)(v, sheetData.formula[v], data.styles);
+                            const f = (0, generate_formula_cell_1.generateCellRowCol)(v, formulaSheetObj[v], data.styles);
                             if (!rF[f.row]) {
                                 rF[f.row] = f.cell;
                             }
@@ -923,6 +755,7 @@ function generateExcel(data) {
             sheetNameApp += "<vt:lpstr>" + ("sheet" + (index + 1)) + "</vt:lpstr>";
             selectedAdded = selectedAdded || !!sheetData.selected;
             const filterMode = sheetData.sortAndfilter ? 'filterMode="1"' : "";
+            mergesCellArray = [...new Set(mergesCellArray)];
             mapData["sheet" + (index + 1)] = {
                 indexId: indexId + 1,
                 key: "sheet" + (index + 1),
@@ -945,10 +778,10 @@ function generateExcel(data) {
                             '" ');
                     }, "<sheetProtection ") + "/>"
                     : "",
-                merges: sheetData.merges
-                    ? sheetData.merges.reduce((mResult, currRef) => {
-                        return (mResult += ' <mergeCell ref="' + currRef + '" />');
-                    }, '<mergeCells count="' + sheetData.merges.length + '">') +
+                merges: mergesCellArray.length > 0
+                    ? mergesCellArray.reduce((mResult, currRef) => {
+                        return mResult + ' <mergeCell ref="' + currRef + '" />';
+                    }, '<mergeCells count="' + mergesCellArray.length + '">') +
                         " </mergeCells>"
                     : "",
                 selectedView: sheetData.selected