+++ dist2/src/utils/content-generator/const-data.d.ts
@@ -1,0 +1,4 @@
+import { FormatMap } from "../../data-model/excel-table";
+export declare const formatMap: FormatMap;
+export declare const cols: string[];
+//# sourceMappingURL=const-data.d.ts.map