+++ dist2/src/utils/multi-value.js
@@ -3,7 +3,18 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.generateMultiStyleValue = void 0;
 function splitAndMatching(v, val, text, textSplited, splitValue, matchValue, styleMatchValue, multiMode) {
     if (!textSplited) {
-        const matchV = text.match(v);
+        let matchV;
+        try {
+            matchV = text.match(v);
+        }
+        catch (error) {
+            if (typeof v == "string") {
+                matchV = text.match("\\" + v);
+            }
+            else {
+                throw error;
+            }
+        }
         if (matchV) {
             if (!multiMode) {
                 matchValue.push(v.toString());
@@ -35,7 +46,18 @@ function splitAndMatching(v, val, text, textSplited, splitValue, matchValue, sty
         let newStyleValue = [];
         const mLength = matchValue.length;
         splitValue.forEach((sp, index) => {
-            const match = sp.match(v);
+            let match;
+            try {
+                match = sp.match(v);
+            }
+            catch (error) {
+                if (typeof v == "string") {
+                    match = sp.match("\\" + v);
+                }
+                else {
+                    throw error;
+                }
+            }
             if (match) {
                 if (!multiMode) {
                     const splitV = sp.split(v).reduce((res, curr, index) => {
@@ -79,7 +101,7 @@ function splitAndMatching(v, val, text, textSplited, splitValue, matchValue, sty
         styleMatchValue,
     };
 }
-function generateMultiStyleValue(multiStyle, text, styles) {
+function generateMultiStyleValue(multiStyle, text, styles, defStyleId) {
     if (typeof multiStyle == "object") {
         let result = "";
         let matchValue = [];
@@ -117,12 +139,14 @@ function generateMultiStyleValue(multiStyle, text, styles) {
             }
         }
         const length = splitValue.length - 1;
+        const elementStyle = defStyleId in styles ? styles[defStyleId] : "";
         for (let index = 0; index < length; index++) {
             const element = splitValue[index];
             const matchElement = matchValue[index];
             const styleID = styleMatchValue[index];
             if (element.length > 0) {
                 result += `<r>
+        ${elementStyle}
             <t xml:space="preserve" >${element}</t>
         </r>`;
             }
@@ -138,6 +162,7 @@ function generateMultiStyleValue(multiStyle, text, styles) {
         result = `<si>
                     ${result}
                     <r>
+        ${elementStyle}
             <t>${splitValue[length]}</t>
         </r>
                     </si>`;